Refact solution structure. Added console project.
Changed the original Console project to a Class Library
Added two new projects
- ReplayParser.Console (containing the Program.cs file of the original
console program)
- ReplayParser.FeatureExtraction

The latter is another console program taking two commandline arguments -
a directory path (not filepath) and a integer value.
The directory path is the path to a folder containing replay files.
The integer value is optional and is a limit to the number of files to
be processed. (default = infinite)